tests/webgl/conformance/limits/gl-max-texture-dimensions.html

<!--

/*
** Copyright (c) 2012 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebGL the max advertized texture size is supported.</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../js/js-test-pre.js"></script>
<script src="../../js/webgl-test-utils.js"> </script>
</head>
<body>
<canvas id="example" width="4" height="4" style="width: 40px; height: 30px;"></canvas>
<div id="description"></div>
<div id="console"></div>
<script id="vshader" type="x-shader/x-vertex">
attribute vec4 vPosition;
attribute vec2 texCoord0;
varying vec2 texCoord;
void main()
{
    gl_Position = vPosition;
    texCoord = texCoord0;
}
</script>

<script id="fshader" type="x-shader/x-fragment">
precision mediump float;
uniform samplerCube tex;
varying vec2 texCoord;
void main()
{
    gl_FragColor = textureCube(tex, normalize(vec3(texCoord, 1)));
}
</script>
<script>
"use strict";
description(document.title);
var wtu = WebGLTestUtils;
var gl = wtu.create3DContext("example");
var program = wtu.setupTexturedQuad(gl);

function shouldBePowerOfTwo(n, name) {
    var power = Math.round(Math.log(n) / Math.log(2));
    if (Math.pow(2, power) == n) {
        testPassed(name + ' is a power of two.');
    } else {
        testFailed(name + ' should be a power of two, but was ' + n);
    }
}

// Note: It seems like a reasonable assuption that a 1xN texture size should
// work. Even 1 by 128k is only 512k
var maxSize = gl.getParameter(gl.MAX_TEXTURE_SIZE);
debug("advertised max size: " + maxSize);
debug("verifying max size is power-of-two (implied by GLES 2.0 section 3.7.1)");
shouldBePowerOfTwo(maxSize, 'Max size');
var testSize = Math.min(maxSize, 128 * 1024);
var pixels = new Uint8Array(testSize * 4);
for (var ii = 0; ii < testSize; ++ii) {
  var off = ii * 4;
  pixels[off + 0] = 0;
  pixels[off + 1] = 255;
  pixels[off + 2] = 128;
  pixels[off + 3] = 255;
}
var tex = gl.createTexture();
gl.bindTexture(gl.TEXTURE_2D, tex);

debug("test " + testSize + "x1");
gl.texImage2D(
    gl.TEXTURE_2D, 0, gl.RGBA, testSize, 1, 0, gl.RGBA, gl.UNSIGNED_BYTE,
    pixels);
gl.generateMipmap(gl.TEXTURE_2D);

wtu.clearAndDrawUnitQuad(gl);
wtu.checkCanvas(gl, [0, 255, 128, 255],
                "Should be 0, 255, 128, 255");
debug("test 1x" + testSize);
gl.texImage2D(
    gl.TEXTURE_2D, 0, gl.RGBA, 1, testSize, 0, gl.RGBA, gl.UNSIGNED_BYTE,
    pixels);
gl.generateMipmap(gl.TEXTURE_2D);

wtu.clearAndDrawUnitQuad(gl);
wtu.checkCanvas(gl, [0, 255, 128, 255],
                "Should be 0, 255, 128, 255");

var program = wtu.setupProgram(
    gl, ['vshader', 'fshader'], ['vPosition', 'texCoord0'], [0, 1]);

wtu.glErrorShouldBe(gl, gl.NO_ERROR, "Should be no errors.");

// NOTE: We can't easily test cube maps because they require width == height
// and we might not have enough memory for maxSize by maxSize texture.

var successfullyParsed = true;

</script>
<script src="../../js/js-test-post.js"></script>

</body>
</html>
